+++ R/dataset_documentation.R
@@ -5,9 +5,9 @@
 #' The example data set includes two data items:
 #'
 #' \itemize{
-#'  \item exampleshrubdata - Data frame with shrub plot data (43 observations and 4 variables).
+#'  \item example_shrubdata_ifn - Data frame with shrub plot data (43 observations and 4 variables).
 #'        Rows are species records and columns are 'ID', 'Species', 'FCC' and 'H'.
-#'  \item exampletreedata - Data frame with tree plot data (118 observations and 11 variables).
+#'  \item example_treedata_ifn - Data frame with tree plot data (118 observations and 11 variables).
 #'    \itemize{
 #'      \item ID - Plot ID
 #'      \item Type - Plot type
@@ -21,10 +21,16 @@
 #'      \item OIF3 - Tree code at the Third Spanish Forest Inventory
 #'      \item FC - Code for the equation describing volume
 #'    }
+#'  \item poblet_trees - Data frame with example tree plot data from Poblet (Catalonia).
+#'    \itemize{
+#'      \item Plot.Code - Plot ID
+#'      \item Indv.Ref - Tree individual (integer)
+#'      \item Species - Species name
+#'      \item Diameter.cm - Tree diameter at breast height (cm)
+#'    }
 #' }
-#'
-#' @name exampledata
-#' @aliases exampleshrubdata exampletreedata
+#' @name example_datasets
+#' @aliases example_shrubdata_ifn example_treedata_ifn poblet_trees
 #' @docType data
 #' @author Toy datasets extracted from Spanish National Forest Inventory.
 #' @keywords data