+++ R/welcome.R
@@ -1,0 +1,6 @@
+.onAttach <- function(lib, pkg)  {
+  packageStartupMessage("Package 'medfateutils' [ver. ",
+                        utils::packageDescription("medfateutils",
+                                                  fields="Version"),"]",
+                        appendLF = TRUE)
+}