+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 0.1.3
 - Keep num orden from IFN data
+- Data set Poblet tree data added
 
 # Version 0.1.2
 - SoilGrids REST API relocated to https://rest.isric.org/soilgrids/v2.0/docs