+++ NEWS.md
@@ -6,6 +6,7 @@
 - Keep num orden from IFN data
 - Data set Poblet tree data added
 - New function modifySoilParams()
+- Update of function 'forest_mapShrubTable'
 
 # Version 0.1.2
 - SoilGrids REST API relocated to https://rest.isric.org/soilgrids/v2.0/docs