+++ NEWS.md
@@ -2,11 +2,14 @@
  NEWS for R Package "medfateutils"
 -----------------------------------
 
+# Version 0.1.4
+- Names changed for IFN example datasets
+- Update of function 'forest_mapShrubTable'
+
 # Version 0.1.3
 - Keep num orden from IFN data
 - Data set Poblet tree data added
 - New function modifySoilParams()
-- Update of function 'forest_mapShrubTable'
 
 # Version 0.1.2
 - SoilGrids REST API relocated to https://rest.isric.org/soilgrids/v2.0/docs